moved menu_bar into dedicated class and fixed some function names
