+++ tagstudio/src/qt/modals/drop_import.py
@@ -182,7 +182,7 @@ class DropImportModal(QWidget):
         pw.from_iterable_function(
             self.copy_files,
             displayed_text,
-            self.driver.add_new_files_callback,
+            self.driver.refresh_directories,
             self.deleteLater,
         )
